+++ legacy/admin/volunteers_control.php
@@ -4,6 +4,7 @@
             <button id="add-volunteer" class="btn btn-sm btn-primary volunteer-manage-buttons" type="button" onclick="includeVolunteers();"><?php echo word('add_volunteer')?></button>
             <button id="save-volunteers" class="btn btn-sm btn-success volunteer-manage-buttons" type="button" onclick="saveVolunteers('<?php echo $data_type?>');"><?php echo word('save_volunteers')?></button>
             <button id="include-group" class="btn btn-sm btn-warning volunteer-manage-buttons" type="button" onclick="showGroupsModal();" style="display: none;"><?php echo word('include_group')?></button>
+            <button id="volunteers-download-list" class="btn btn-sm btn-secondary volunteer-manage-buttons" type="button" style="display: none;">Download Volunteer List</button>
         </div>
     </div>
 </div>